+++ Scripts/HTMLDocsAutomation/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## [Unreleased]
 
 
+## [19.10.0] - 2019-10-03
+Added support for commands that have only headlines and no text in the human readable output.
+
 ## [19.9.1] - 2019-09-18
  - Added the *permissions* argument with the following options:
     - **per-command** - the permissions entry will be displayed in every command section.